requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
altair==5.2.0
anyio==3.7.1
appnope==0.1.3
argilla==1.19.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.1.0
backoff==2.2.1
bcrypt==4.1.1
bertopic==0.16.0
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.0
Cython==0.29.36
datasets==2.15.0
debugpy==1.8.0
decorator==5.1.1
Deprecated==1.2.14
dill==0.3.7
diskcache==5.6.3
et-xmlfile==1.1.0
evaluate==0.4.1
exceptiongroup==1.2.0
executing==2.0.1
fastapi==0.104.1
fastjsonschema==2.19.0
filelock==3.13.1
frozenlist==1.4.0
fsspec==2023.10.0
gitdb==4.0.11
GitPython==3.1.40
h11==0.14.0
hdbscan==0.8.33
httpcore==0.18.0
httptools==0.6.1
httpx==0.25.0
huggingface-hub==0.19.4
idna==3.6
importlib-metadata==6.8.0
ipykernel==6.27.1
ipython==8.18.1
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
jupyter_client==8.6.0
jupyter_core==5.5.0
llama_cpp_python==0.2.2
llvmlite==0.41.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mdurl==0.1.2
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.2.1
nltk==3.8.1
numba==0.58.1
numpy==1.23.5
openpyxl==3.1.2
packaging==23.2
pandas==1.5.3
parso==0.8.3
passlib==1.7.4
pexpect==4.9.0
Pillow==10.1.0
platformdirs==4.0.0
plotly==5.18.0
prompt-toolkit==3.0.41
protobuf==4.25.1
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==14.0.1
pyarrow-hotfix==0.6
pydantic==1.10.13
pydeck==0.8.1b0
Pygments==2.17.2
pynndescent==0.5.11
pypng==0.20220715.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.1
qrcode==7.4.2
referencing==0.31.0
regex==2023.10.3
requests==2.31.0
responses==0.18.0
rich==13.7.0
rpds-py==0.13.1
safetensors==0.4.1
scikit-learn==1.3.2
scipy==1.11.4
sentence-transformers==2.2.2
sentencepiece==0.1.99
setfit==0.7.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
st-execbox==1.0.0
stack-data==0.6.3
starlette==0.27.0
streamlit==1.29.0
streamlit-ace==0.1.1
streamlit-code-editor==0.1.10
streamlit-reveal-slides==0.2.0
sympy==1.12
tenacity==8.2.3
/identify_categories/libs
threadpoolctl==3.2.0
tokenizers==0.15.0
toml==0.10.2
toolz==0.12.0
torch==2.1.1
torchvision==0.16.1
tornado==6.3.3
tqdm==4.66.1
traitlets==5.14.0
transformers==4.35.2
typer==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
umap-learn==0.5.5
urllib3==2.1.0
uvicorn==0.24.0.post1
uvloop==0.19.0
validators==0.22.0
watchfiles==0.21.0
wcwidth==0.2.12
websockets==12.0
wrapt==1.14.1
xxhash==3.4.1
yarl==1.9.3
zipp==3.17.0
